Feature/simplify storage typing

+++ backend/src/main/java/com/bakdata/conquery/models/config/XodusStoreFactory.java
@@ -1,14 +1,12 @@
 package com.bakdata.conquery.models.config;
 
-import static com.bakdata.conquery.io.storage.StoreInfo.*;
+import static com.bakdata.conquery.io.storage.StoreMappings.*;
 
 import java.io.File;
 import java.io.IOException;
 import java.nio.file.Path;
-import java.util.ArrayList;
 import java.util.Collection;
-import java.util.List;
-import java.util.Optional;
+import java.util.Objects;
 import java.util.Queue;
 import java.util.concurrent.ConcurrentLinkedQueue;
 import java.util.concurrent.ExecutorService;
@@ -16,6 +14,7 @@ import java.util.concurrent.Executors;
 import java.util.concurrent.TimeUnit;
 import java.util.function.Function;
 
+import javax.annotation.Nullable;
 import javax.validation.Valid;
 import javax.validation.Validator;
 import javax.validation.constraints.Min;
@@ -30,12 +29,13 @@ import com.bakdata.conquery.io.storage.IdentifiableStore;
 import com.bakdata.conquery.io.storage.NamespaceStorage;
 import com.bakdata.conquery.io.storage.NamespacedStorage;
 import com.bakdata.conquery.io.storage.Store;
-import com.bakdata.conquery.io.storage.StoreInfo;
+import com.bakdata.conquery.io.storage.StoreMappings;
 import com.bakdata.conquery.io.storage.WorkerStorage;
 import com.bakdata.conquery.io.storage.xodus.stores.BigStore;
 import com.bakdata.conquery.io.storage.xodus.stores.CachedStore;
 import com.bakdata.conquery.io.storage.xodus.stores.SerializingStore;
 import com.bakdata.conquery.io.storage.xodus.stores.SingletonStore;
+import com.bakdata.conquery.io.storage.xodus.stores.StoreInfo;
 import com.bakdata.conquery.io.storage.xodus.stores.WeakCachedStore;
 import com.bakdata.conquery.io.storage.xodus.stores.XodusStore;
 import com.bakdata.conquery.models.auth.entities.Group;
@@ -113,7 +113,8 @@ public class XodusStoreFactory implements StoreFactory {
 	/**
 	 * When set, all values that could not be deserialized from the persistent store, are dump into individual files.
 	 */
-	private Optional<File> unreadableDataDumpDirectory = Optional.empty();
+	@Nullable
+	private File unreadableDataDumpDirectory = null;
 
 	@JsonIgnore
 	private transient Validator validator;
@@ -122,7 +123,7 @@ public class XodusStoreFactory implements StoreFactory {
 	private transient ObjectMapper objectMapper = Jackson.BINARY_MAPPER.copy();
 
 	@JsonIgnore
-	private BiMap<File, Environment> activeEnvironments = HashBiMap.create();
+	private final BiMap<File, Environment> activeEnvironments = HashBiMap.create();
 
 	@JsonIgnore
 	private final transient Multimap<Environment, jetbrains.exodus.env.Store>
@@ -172,7 +173,7 @@ public class XodusStoreFactory implements StoreFactory {
 		ExecutorService loaders = Executors.newFixedThreadPool(getNThreads());
 
 
-		for (File directory : baseDir.listFiles((file, name) -> file.isDirectory() && name.startsWith(prefix))) {
+		for (File directory : Objects.requireNonNull(baseDir.listFiles((file, name) -> file.isDirectory() && name.startsWith(prefix)))) {
 
 			final String name = directory.getName();
 
@@ -213,19 +214,9 @@ public class XodusStoreFactory implements StoreFactory {
 		return storages;
 	}
 
-	private List<String> getRelativePathElements(Path path) {
-		ArrayList<String> list = new ArrayList<>();
-		Path relative = getDirectory().relativize(path);
-		for (int i = 0; i < relative.getNameCount(); i++) {
-			list.add(relative.getName(i).toString());
-		}
-		return list;
-	}
-
 	private boolean environmentHasStores(File pathName) {
 		Environment env = findEnvironment(pathName);
-		boolean exists = env.computeInTransaction(t -> env.storeExists(StoreInfo.DATASET.getName(), t));
-		env.computeInTransaction(t -> env.getAllStoreNames(t));
+		boolean exists = env.computeInTransaction(t -> env.storeExists(StoreMappings.DATASET.storeInfo().getName(), t));
 		if (!exists) {
 			closeEnvironment(env);
 		}
@@ -234,17 +225,17 @@ public class XodusStoreFactory implements StoreFactory {
 
 	@Override
 	public SingletonStore<Dataset> createDatasetStore(String pathName) {
-		return DATASET.singleton(createStore(findEnvironment(pathName), validator, DATASET));
+		return StoreMappings.singleton(createStore(findEnvironment(pathName), validator, DATASET));
 	}
 
 	@Override
 	public IdentifiableStore<SecondaryIdDescription> createSecondaryIdDescriptionStore(CentralRegistry centralRegistry, String pathName) {
-		return SECONDARY_IDS.identifiable(createStore(findEnvironment(pathName), validator, SECONDARY_IDS), centralRegistry);
+		return StoreMappings.identifiable(createStore(findEnvironment(pathName), validator, SECONDARY_IDS), centralRegistry);
 	}
 
 	@Override
 	public IdentifiableStore<Table> createTableStore(CentralRegistry centralRegistry, String pathName) {
-		return TABLES.identifiable(createStore(findEnvironment(pathName), validator, TABLES), centralRegistry);
+		return StoreMappings.identifiable(createStore(findEnvironment(pathName), validator, TABLES), centralRegistry);
 	}
 
 	@Override
@@ -257,46 +248,45 @@ public class XodusStoreFactory implements StoreFactory {
 
 		synchronized (openStoresInEnv) {
 			bigStore =
-					new BigStore<>(this, validator, environment, DICTIONARIES, openStoresInEnv.get(environment), this::closeEnvironment, this::removeEnvironment, namespaceCollection
-																																										  .injectInto(objectMapper));
+					new BigStore<>(
+							this,
+							validator,
+							environment,
+							DICTIONARIES.storeInfo(),
+							this::closeStore,
+							this::removeStore,
+							namespaceCollection.injectInto(objectMapper));
 		}
 
-		final Store<IId<Dictionary>, Dictionary> result;
-
-		// TODO this looks like dictionaries are double cached
 		if (useWeakDictionaryCaching) {
-			result = new WeakCachedStore<>(bigStore, getWeakCacheDuration());
+			return StoreMappings.identifiableCachedStore(new WeakCachedStore<>(bigStore, getWeakCacheDuration()), centralRegistry);
 		}
-		else {
-			result = DICTIONARIES.cached(bigStore);
-		}
-
-		return DICTIONARIES.identifiableCachedStore(result, centralRegistry);
+		return StoreMappings.identifiable(bigStore, centralRegistry);
 	}
 
 	@Override
 	public IdentifiableStore<Concept<?>> createConceptStore(CentralRegistry centralRegistry, String pathName) {
-		return CONCEPTS.identifiable(createStore(findEnvironment(pathName), validator, CONCEPTS), centralRegistry);
+		return StoreMappings.identifiable(createStore(findEnvironment(pathName), validator, CONCEPTS), centralRegistry);
 	}
 
 	@Override
 	public IdentifiableStore<Import> createImportStore(CentralRegistry centralRegistry, String pathName) {
-		return IMPORTS.identifiable(createStore(findEnvironment(pathName), validator, IMPORTS), centralRegistry);
+		return StoreMappings.identifiable(createStore(findEnvironment(pathName), validator, IMPORTS), centralRegistry);
 	}
 
 	@Override
 	public IdentifiableStore<CBlock> createCBlockStore(CentralRegistry centralRegistry, String pathName) {
-		return C_BLOCKS.identifiable(createStore(findEnvironment(pathName), validator, C_BLOCKS), centralRegistry);
+		return StoreMappings.identifiable(createStore(findEnvironment(pathName), validator, C_BLOCKS), centralRegistry);
 	}
 
 	@Override
 	public IdentifiableStore<Bucket> createBucketStore(CentralRegistry centralRegistry, String pathName) {
-		return BUCKETS.identifiable(createStore(findEnvironment(pathName), validator, BUCKETS), centralRegistry);
+		return StoreMappings.identifiable(createStore(findEnvironment(pathName), validator, BUCKETS), centralRegistry);
 	}
 
 	@Override
 	public SingletonStore<WorkerInformation> createWorkerInformationStore(String pathName) {
-		return WORKER.singleton(createStore(findEnvironment(pathName), validator, WORKER));
+		return StoreMappings.singleton(createStore(findEnvironment(pathName), validator, WORKER));
 	}
 
 	@Override
@@ -305,7 +295,7 @@ public class XodusStoreFactory implements StoreFactory {
 
 		synchronized (openStoresInEnv) {
 			final BigStore<Boolean, EntityIdMap> bigStore =
-					new BigStore<>(this, validator, environment, ID_MAPPING, openStoresInEnv.get(environment), this::closeEnvironment, this::removeEnvironment, objectMapper);
+					new BigStore<>(this, validator, environment, ID_MAPPING.storeInfo(), this::closeStore, this::removeStore, objectMapper);
 
 			return new SingletonStore<>(new CachedStore<>(bigStore));
 		}
@@ -313,43 +303,43 @@ public class XodusStoreFactory implements StoreFactory {
 
 	@Override
 	public SingletonStore<WorkerToBucketsMap> createWorkerToBucketsStore(String pathName) {
-		return WORKER_TO_BUCKETS.singleton(createStore(findEnvironment(pathName), validator, WORKER_TO_BUCKETS));
+		return StoreMappings.singleton(createStore(findEnvironment(pathName), validator, WORKER_TO_BUCKETS));
 	}
 
 	@Override
 	public SingletonStore<StructureNode[]> createStructureStore(String pathName, SingletonNamespaceCollection centralRegistry) {
-		return STRUCTURE.singleton(createStore(findEnvironment(pathName), validator, STRUCTURE), centralRegistry);
+		return StoreMappings.singleton(createStore(findEnvironment(pathName), validator, STRUCTURE), centralRegistry);
 	}
 
 	@Override
 	public IdentifiableStore<ManagedExecution<?>> createExecutionsStore(CentralRegistry centralRegistry, DatasetRegistry datasetRegistry, String pathName) {
-		return EXECUTIONS.identifiable(createStore(findEnvironment(resolveSubDir(pathName, "executions")), validator, EXECUTIONS), centralRegistry, datasetRegistry);
+		return StoreMappings.identifiable(createStore(findEnvironment(resolveSubDir(pathName, "executions")), validator, EXECUTIONS), centralRegistry, datasetRegistry);
 	}
 
 	@Override
 	public IdentifiableStore<FormConfig> createFormConfigStore(CentralRegistry centralRegistry, DatasetRegistry datasetRegistry, String pathName) {
-		return FORM_CONFIG.identifiable(createStore(findEnvironment(resolveSubDir(pathName, "formConfigs")), validator, FORM_CONFIG), centralRegistry, datasetRegistry);
+		return StoreMappings.identifiable(createStore(findEnvironment(resolveSubDir(pathName, "formConfigs")), validator, FORM_CONFIG), centralRegistry, datasetRegistry);
 	}
 
 	@Override
 	public IdentifiableStore<User> createUserStore(CentralRegistry centralRegistry, String pathName) {
-		return AUTH_USER.identifiable(createStore(findEnvironment(resolveSubDir(pathName, "users")), validator, AUTH_USER), centralRegistry);
+		return StoreMappings.identifiable(createStore(findEnvironment(resolveSubDir(pathName, "users")), validator, AUTH_USER), centralRegistry);
 	}
 
 	@Override
 	public IdentifiableStore<Role> createRoleStore(CentralRegistry centralRegistry, String pathName) {
-		return AUTH_ROLE.identifiable(createStore(findEnvironment(resolveSubDir(pathName, "roles")), validator, AUTH_ROLE), centralRegistry);
+		return StoreMappings.identifiable(createStore(findEnvironment(resolveSubDir(pathName, "roles")), validator, AUTH_ROLE), centralRegistry);
 	}
 
 
 	@Override
 	public IdentifiableStore<Group> createGroupStore(CentralRegistry centralRegistry, String pathName) {
-		return AUTH_GROUP.identifiable(createStore(findEnvironment(resolveSubDir(pathName, "groups")), validator, AUTH_GROUP), centralRegistry);
+		return StoreMappings.identifiable(createStore(findEnvironment(resolveSubDir(pathName, "groups")), validator, AUTH_GROUP), centralRegistry);
 	}
 
 	@Override
 	public SingletonStore<Dictionary> createPrimaryDictionaryStore(String pathName, SingletonNamespaceCollection namespaceCollection) {
-		return PRIMARY_DICTIONARY.singleton(createStore(findEnvironment(pathName), validator, PRIMARY_DICTIONARY), namespaceCollection);
+		return StoreMappings.singleton(createStore(findEnvironment(pathName), validator, PRIMARY_DICTIONARY), namespaceCollection);
 	}
 
 	private File resolveSubDir(String... subdirs) {
@@ -384,11 +374,22 @@ public class XodusStoreFactory implements StoreFactory {
 		}
 	}
 
+	private void closeStore(jetbrains.exodus.env.Store store) {
+		Environment env = store.getEnvironment();
+		Collection<jetbrains.exodus.env.Store> stores = openStoresInEnv.get(env);
+		stores.remove(store);
+		log.info("Closed XodusStore: {}", this);
+
+		if (!stores.isEmpty()) {
+			return;
+		}
+		log.info("Closed last XodusStore in Environment. Closing Environment as well: {}", env.getLocation());
+
+		closeEnvironment(env);
+	}
+
 	private void closeEnvironment(Environment env) {
 		synchronized (activeEnvironments) {
-			if (env == null) {
-				return;
-			}
 
 			if (activeEnvironments.remove(activeEnvironments.inverse().get(env)) == null) {
 				return;
@@ -397,25 +398,35 @@ public class XodusStoreFactory implements StoreFactory {
 		}
 	}
 
-	private void removeEnvironment(Environment env) {
-		log.info("Deleting Environment[{}]", env.getLocation());
+	private void removeStore(jetbrains.exodus.env.Store store) {
+		Environment env = store.getEnvironment();
+		Collection<jetbrains.exodus.env.Store> stores = openStoresInEnv.get(env);
+
+		if (!stores.isEmpty()) {
+			return;
+		}
+
+		log.info("Removed last XodusStore in Environment. Removing Environment as well: {}", env.getLocation());
+		env.close();
+
 		try {
 			FileUtil.deleteRecursive(Path.of(env.getLocation()));
 		}
 		catch (IOException e) {
-			log.error("Cannot delete directory of removed Environment[{}]", env.getLocation(), log.isDebugEnabled() ? e : null);
+			log.error("Cannot delete directory of removed Environment[{}]", env.getLocation(), e);
 		}
 	}
 
-	public <KEY, VALUE> Store<KEY, VALUE> createStore(Environment environment, Validator validator, StoreInfo storeId) {
+	public <KEY, VALUE> Store<KEY, VALUE> createStore(Environment environment, Validator validator, StoreMappings storeId) {
+		final StoreInfo<KEY, VALUE> storeInfo = storeId.storeInfo();
 		synchronized (openStoresInEnv) {
 			return new CachedStore<>(
 					new SerializingStore<>(
-							this,
-							new XodusStore(environment, storeId, openStoresInEnv.get(environment), this::closeEnvironment, this::removeEnvironment),
+							new XodusStore(environment, storeInfo.getName(), this::closeStore, this::removeStore),
 							validator,
-							storeId,
-							objectMapper
+							objectMapper,
+							storeInfo.getKeyType(),
+							storeInfo.getValueType(), this.isRemoveUnreadableFromStore(), this.getUnreadableDataDumpDirectory(), this.isValidateOnWrite()
 					));
 		}
 	}
